move get_doc_ids to dbaccessors
