+++ corehq/apps/cleanup/management/commands/migrate_repeat_records.py
@@ -1,4 +1,4 @@
-from corehq.apps.domain.utils import get_doc_ids
+from corehq.apps.domain.dbaccessors import get_doc_ids
 from dimagi.utils.couch.database import iter_docs
 from django.core.management.base import BaseCommand
 from corehq.apps.domain.models import Domain
