+++ corehq/apps/cleanup/management/commands/repeater_times.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from corehq.apps.domain.utils import get_doc_ids
+from corehq.apps.domain.dbaccessors import get_doc_ids
 from dimagi.utils.couch.database import iter_docs
 from django.core.management.base import BaseCommand
 from corehq.apps.domain.models import Domain
