+++ corehq/apps/domain/utils.py
@@ -1,11 +1,11 @@
 import re
+
 from couchdbkit import ResourceNotFound
 from django.conf import settings
 
 from corehq import toggles
 from corehq.apps.domain.models import Domain
 from corehq.util.quickcache import quickcache
-
 from dimagi.utils.couch.database import get_db
 
 
@@ -65,20 +65,6 @@ def get_dummy_domain(domain_type=None):
     dummy_domain = Domain()
     dummy_domain.commtrack_enabled = (domain_type == 'commtrack')
     return dummy_domain
-
-
-def get_doc_ids(domain, doc_type, database=None):
-    """
-    Given a domain and doc type, get all docs matching that domain and type
-    """
-    if not database:
-        database = get_db()
-    return [row['id'] for row in database.view('domain/docs',
-        startkey=[domain, doc_type],
-        endkey=[domain, doc_type, {}],
-        reduce=False,
-        include_docs=False,
-    )]
 
 
 def user_has_custom_top_menu(domain_name, couch_user):
